added points update

+++ src/main/java/com/logicbyte/chillers/model/Player.java
@@ -1,10 +1,5 @@
 package com.logicbyte.chillers.model;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.Transient;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
 
@@ -22,7 +17,7 @@ import lombok.experimental.SuperBuilder;
 @AllArgsConstructor
 //@JsonInclude(JsonInclude.Include.NON_DEFAULT)
 public class Player {
-    private Long id;
+    private Integer id;
     private String nickname;
     private String image_url;
     private int playerPoints;
